Begins transition to new style and to Vue.js
This commit represents the beginning of the transition to a new style
for the website and also a technical switch from React to Vue.js.

Essentially all previous files have been overwritten by the ones created
by the command `vue init webpack` which created a ready-to-go Vue.js app
configured with Webpack, testing frameworks and routing.