Reset application with new Preact+TS+Parcel setup
With the objective to reboot the personal webiste, a new fresh start is
necessary: current design, usability and technical stack are not
appreciated anymore for this project.

This commit starts fresh with a new setup using Typescript (for all the
good reasons), Parcel (for its simplicity since this is not a complex
project) and Preact (the fanciness of React with much smaller bundle
size, again because this is a simple project).